pyproject.toml
[tool.poetry]
name = "pat"
version = "0.1.0"
description = ""
authors = ["Tristan Kernan <tristanmkernan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-allauth = "^0.54.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-htmx = "^1.14.0"
gunicorn = "^20.1.0"
python-decouple = "^3.8"
whitenoise = "^6.4.0"
django-guardian = "^2.4.0"
django-taggit = "^3.1.0"
django-tables2 = "^2.5.3"
django-filter = "^23.2"
django-extensions = "^3.2.1"
factory-boy = "^3.2.1"
django-storages = "^1.13.2"
boto3 = "^1.28.30"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipdb = "^0.13.13"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 99
select = ["F", "E", "W", "Q",]
exclude = [
    "**/migrations/*.py"
]
ignore = ["F401"]
